+++ NEWS.md
@@ -1,4 +1,4 @@
-# glue (development version)
+# glue 1.5.0
 
 ## Breaking changes